Extract CartViewModel to own file. Remove cart-related methods from dashboard
