+++ WooCommerce/Classes/POS/ViewModels/PointOfSaleDashboardViewModel.swift
@@ -10,48 +10,6 @@ import enum Yosemite.OrderStatusEnum
 import struct Yosemite.POSCartItem
 import struct Yosemite.Order
 
-final class CartViewModel: ObservableObject {
-    @Published private(set) var itemsInCart: [CartItem] = []
-
-    init() { }
-
-    var isEmpty: Bool {
-        itemsInCart.isEmpty
-    }
-
-    var isCartCollapsed: Bool {
-        itemsInCart.isEmpty
-    }
-
-    var itemToScrollToWhenCartUpdated: CartItem? {
-        itemsInCart.last
-    }
-
-    var itemsInCartLabel: String? {
-        switch itemsInCart.count {
-        case 0:
-            return nil
-        default:
-            return String.pluralize(itemsInCart.count,
-                                    singular: "%1$d item",
-                                    plural: "%1$d items")
-        }
-    }
-
-    func addItemToCart(_ item: POSItem) {
-        let cartItem = CartItem(id: UUID(), item: item, quantity: 1)
-        itemsInCart.append(cartItem)
-    }
-
-    func removeItemFromCart(_ cartItem: CartItem) {
-        itemsInCart.removeAll(where: { $0.id == cartItem.id })
-    }
-
-    func removeAllItemsFromCart() {
-        itemsInCart.removeAll()
-    }
-}
-
 final class PointOfSaleDashboardViewModel: ObservableObject {
     enum PaymentState {
         case idle
@@ -166,32 +124,6 @@ final class PointOfSaleDashboardViewModel: ObservableObject {
 
     var canDeleteItemsFromCart: Bool {
         return orderStage != .finalizing
-    }
-
-    var isCartCollapsed: Bool {
-        cartViewModel.isCartCollapsed
-    }
-
-    var itemToScrollToWhenCartUpdated: CartItem? {
-        cartViewModel.itemToScrollToWhenCartUpdated
-    }
-
-    func addItemToCart(_ item: POSItem) {
-        cartViewModel.addItemToCart(item)
-    }
-
-    func removeItemFromCart(_ cartItem: CartItem) {
-        cartViewModel.removeItemFromCart(cartItem)
-        checkIfCartEmpty()
-    }
-
-    func removeAllItemsFromCart() {
-        cartViewModel.removeAllItemsFromCart()
-        checkIfCartEmpty()
-    }
-
-    var itemsInCartLabel: String? {
-        cartViewModel.itemsInCartLabel
     }
 
     private func checkIfCartEmpty() {
